requirements.txt
fastapi
langchain_community
langchain_core
langchain_openai
langchain_text_splitters
pydantic
python-dotenv
Requests
streamlit
uvicorn
faiss-cpu
tiktoken